transientdisk: move flash drive implementation in a separate file
Also encapsulate the entire MemoryMessage handling.
